feat: update `package.json` with `npm version`
