+++ lib/prepare.js
@@ -1,25 +1,37 @@
 const path = require('path');
 const {move} = require('fs-extra');
 const execa = require('execa');
-const updatePackageVersion = require('./update-package-version');
 
 module.exports = async ({tarballDir, pkgRoot}, {cwd, env, stdout, stderr, nextRelease: {version}, logger}) => {
   const basePath = pkgRoot ? path.resolve(cwd, pkgRoot) : cwd;
-  await updatePackageVersion(version, basePath, logger);
+
+  logger.log('Write version %s to package.json in %s', version, basePath);
+
+  const versionResult = execa('npm', ['version', version, '--no-git-tag-version'], {cwd: basePath, env});
+  versionResult.stdout.pipe(
+    stdout,
+    {end: false}
+  );
+  versionResult.stderr.pipe(
+    stderr,
+    {end: false}
+  );
+
+  await versionResult;
 
   if (tarballDir) {
     logger.log('Creating npm package version %s', version);
-    const result = execa('npm', ['pack', basePath], {cwd, env});
-    result.stdout.pipe(
+    const packResult = execa('npm', ['pack', basePath], {cwd, env});
+    packResult.stdout.pipe(
       stdout,
       {end: false}
     );
-    result.stderr.pipe(
+    packResult.stderr.pipe(
       stderr,
       {end: false}
     );
 
-    const tarball = (await result).stdout.split('\n').pop();
+    const tarball = (await packResult).stdout.split('\n').pop();
     await move(path.resolve(cwd, tarball), path.resolve(cwd, tarballDir.trim(), tarball));
   }
 };
