+++ package.json
@@ -18,15 +18,12 @@
   "dependencies": {
     "@semantic-release/error": "^2.2.0",
     "aggregate-error": "^1.0.0",
-    "detect-indent": "^5.0.0",
-    "detect-newline": "^2.1.0",
     "execa": "^1.0.0",
     "fs-extra": "^7.0.0",
     "lodash": "^4.17.4",
     "nerf-dart": "^1.0.0",
     "normalize-url": "^4.0.0",
     "npm": "^6.3.0",
-    "parse-json": "^4.0.0",
     "rc": "^1.2.8",
     "read-pkg": "^4.0.0",
     "registry-auth-token": "^3.3.1"
